2commit

+++ Program.cs
@@ -1,3 +1,3 @@
-﻿int numberA = 12;
-int numberB = 5;
+﻿double numberA = 12;
+double numberB = 5;
 Console.WriteLine (numberA / numberB);
